+++ src/hint_processor/builtin_hint_processor/secp/ec_utils.rs
@@ -349,8 +349,7 @@ pub fn fast_ec_add_assign_new_x(
         point0_y_d2.as_ref(),
     );
 
-    let value = (&slope * &slope - &x0 - x1).mod_floor(&secp_p);
-
+    let value = (&slope * &slope - &x0 - &x1).mod_floor(&secp_p);
     //Assign variables to vm scope
     exec_scopes.insert_value("slope", slope);
     exec_scopes.insert_value("x0", x0);